memecoin_tracer/src/requirements.txt
tweepy==4.10.0
requests==2.28.1
textblob==0.17.1
matplotlib==3.5.2
praw==7.6.0
python-telegram-bot==13.15
backoff==2.2.1